requirements.txt
openai
supabase
python-dotenv 